ZSH: custom part refactoring

+++ .zsh/custom/alias.zsh
@@ -1,20 +1,29 @@
 # Aliases block
 alias tailf='tail -f'
 alias exit='sync; sync; sync; clear; exit'
-alias vi='/usr/local/bin/nvim'
-alias vim='/usr/local/bin/nvim'
 
-# MacOs aliases
-alias resetsound="ps aux | grep 'coreaudio[d]' | awk '{print \$2}' | xargs sudo kill"
-alias resetdns='sudo killall -HUP mDNSResponder && sudo pkill dnsmasq && echo DNS cleared'
-# alias gitk='open -a "SmartGit" --args ${PWD}/'
-alias gitk='open -a SourceTree --args "${PWD}/"'
-alias git='LANGUAGE=en_US.UTF-8 git'
+# Neovim
+NVIM=$(which nvim)
+alias vi="$NVIM"
+alias vim="$NVIM"
 
 # fzf
 alias cdf="cd \$(dirname \"\$(fzf)\")"
 
-alias idea='open -a "IntelliJ IDEA CE" .'
+# GIT
+alias git='LANGUAGE=en_US.UTF-8 git'
+
+# Idea
+if [[ "$OSTYPE" == "darwin"* ]]; then
+
+    # MacOs aliases
+    alias resetsound="ps aux | grep 'coreaudio[d]' | awk '{print \$2}' | xargs sudo kill"
+    alias resetdns='sudo killall -HUP mDNSResponder && sudo pkill dnsmasq && echo DNS cleared'
+    # alias gitk='open -a "SmartGit" --args ${PWD}/'
+    alias gitk='open -a SourceTree --args "${PWD}/"'
+
+    alias idea='open -a "IntelliJ IDEA CE" .'
+fi
 
 #k8s
 alias k='kubectl'
